requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
audio-recorder-streamlit==0.0.8
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
elevenlabs==0.2.27
exceptiongroup==1.2.0
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.1
google-cloud==0.34.0
google-cloud-speech==2.25.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.26.4
openai==1.12.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
stack-data==0.6.3
streamlit==1.31.1
streamlit-audiorec==0.1.3
streamlit-option-menu==0.3.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
websockets==12.0
zipp==3.17.0
